server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
cryptography==41.0.7
pycryptodome==3.19.0
numpy==1.24.3
gunicorn==21.2.0